Moved CLI into own class FrameworkToolboxCli
